CNS-477: fix comments, move functions, and more small changes
