+++ proto/plans/plan.proto
@@ -47,8 +47,9 @@ message ChainPolicy {
     repeated string apis = 2 [(gogoproto.nullable) = false, (gogoproto.moretags) = "mapstructure:\"apis\""];
 }
 
-// The geolocations' values are powers of 2 that act as a bit map.
-// To add a new geolocation, put it before GL and assign it with the next power of 2.
+// The geolocation values are encoded as bits in a bitmask, with two special values:
+// GLS is set to 0 so it will be restrictive with the AND operator.
+// GL is set to -1 so it will be permissive with the AND operator.
 enum Geolocation {
     GLS = 0; // Global-strict
     USC = 0x1; // US-Center
