+++ x/pairing/client/cli/tx_stake_provider.go
@@ -10,9 +10,11 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	commontypes "github.com/lavanet/lava/common/types"
 	"github.com/lavanet/lava/utils"
 	epochstoragetypes "github.com/lavanet/lava/x/epochstorage/types"
 	"github.com/lavanet/lava/x/pairing/types"
+	planstypes "github.com/lavanet/lava/x/plans/types"
 	spectypes "github.com/lavanet/lava/x/spec/types"
 	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
@@ -32,7 +34,7 @@ func CmdStakeProvider() *cobra.Command {
 		[chain-id] is the spec the provider wishes to support
 		[amount] is the ulava amount to be staked
 		[endpoint endpoint ...] are a space separated list of HOST:PORT,useType,geolocation, should be defined within "quotes". Note that you can specify the geolocation using a single uint64 (which will be treated as a bitmask (see plan.proto)) or using one of the geolocation codes. Valid geolocation codes: AF, AS, AU, EU, USE (US east), USC, USW, GL (global).
-		[geolocation] should be the geolocation code to be staked for. You can also use the geolocation codes syntax: EU,AU,AF. Note that this geolocation should be a union of the endpoints' geolocations.`,
+		[geolocation] should be the geolocation codes to be staked for. You can also use the geolocation codes syntax: EU,AU,AF. Note that this geolocation should be a union of the endpoints' geolocations.`,
 		Example: `
 		lavad tx pairing stake-provider "ETH1" 500000ulava "my-provider.com/rpc,jsonrpc,1" 1 -y --from provider-wallet --provider-moniker "my-moniker" --gas-adjustment "1.5" --gas "auto" --gas-prices $GASPRICE
 
@@ -45,7 +47,7 @@ func CmdStakeProvider() *cobra.Command {
 				return err
 			}
 
-			argEndpoints, argGeolocation, err := types.HandleEndpointsAndGeolocationArgs(strings.Fields(args[2]), args[3])
+			argEndpoints, argGeolocation, err := HandleEndpointsAndGeolocationArgs(strings.Fields(args[2]), args[3])
 			if err != nil {
 				return err
 			}
@@ -196,4 +198,68 @@ func CmdBulkStakeProvider() *cobra.Command {
 	flags.AddTxFlagsToCmd(cmd)
 
 	return cmd
+}
+
+func HandleEndpointsAndGeolocationArgs(endpArg []string, geoArg string) (endp []epochstoragetypes.Endpoint, geo uint64, err error) {
+	for _, endpointStr := range endpArg {
+		splitted := strings.Split(endpointStr, ",")
+		if len(splitted) != 3 {
+			return nil, 0, fmt.Errorf("invalid argument format in endpoints, received: %s", endpointStr)
+		}
+
+		// geolocation of an endpoint can be uint or a string representing a single geo region
+		var geoloc uint64
+		geoloc, valid := types.IsValidGeoEnum(splitted[2])
+		if !valid {
+			geoloc, err = strconv.ParseUint(splitted[2], 10, 64)
+			if err != nil {
+				return nil, 0, fmt.Errorf("invalid argument format in endpoints, geolocation must be a number or valid geolocation string")
+			}
+		}
+
+		// if the user specified global ("GL"), append the endpoint in all possible geolocations
+		if geoloc == uint64(planstypes.Geolocation_GL) {
+			for geoName, geoVal := range planstypes.Geolocation_value {
+				if geoName == planstypes.Geolocation_GL.String() || geoName == planstypes.Geolocation_GLS.String() {
+					continue
+				}
+				endpoint := epochstoragetypes.Endpoint{IPPORT: splitted[0], UseType: splitted[1], Geolocation: uint64(geoVal)}
+				endp = append(endp, endpoint)
+			}
+		} else {
+			// if it's not global, verify that the endpoint's geolocation represents a single geo region
+			geoRegions, _ := types.ExtractGeolocations(geoloc)
+			if len(geoRegions) != 1 {
+				return nil, 0, fmt.Errorf("invalid geolocation for endpoint, must represent one region")
+			}
+			endpoint := epochstoragetypes.Endpoint{IPPORT: splitted[0], UseType: splitted[1], Geolocation: geoloc}
+			endp = append(endp, endpoint)
+		}
+	}
+
+	// handle the case of string geolocation (example: "EU,AF,AS")
+	splitted := strings.Split(geoArg, ",")
+	strEnums := commontypes.RemoveDuplicatesFromSlice(splitted)
+	for _, s := range strEnums {
+		g, valid := types.IsValidGeoEnum(s)
+		if valid {
+			// if one of the endpoints is global, assign global value and break
+			if g == uint64(planstypes.Geolocation_GL) {
+				geo = types.GetCurrentGlobalGeolocation()
+				break
+			}
+			// for non-global geolocations constructs the final geolocation uint value (addition works because we
+			// removed duplicates and the geo regions represent a bitmap)
+			geo += g
+		}
+	}
+
+	// geolocation is not a list of enums, try to parse it as an uint
+	if geo == 0 {
+		geo, err = cast.ToUint64E(geoArg)
+		if err != nil {
+			return nil, 0, err
+		}
+	}
+	return endp, geo, nil
 }
